requirements.txt
baostock
dolphindb
apscheduler
chinesecalendar
pydantic-settings
pydantic
tenacity
pandas
pyyaml
pyinstaller
requests
mplfinance
matplotlib
